fix(home): separate client component for translations in home page to fix async component error
The home page server component now only handles runtime feature flags and redirects, and renders the new client component for UI and translation logic.\n\nThis avoids the Next-Intl error: 'useTranslations is not callable within an async component'. See https://next-intl-docs.vercel.app/docs/environments/server-client-components#async-components for details.\n\nNecessary for deployment on Vercel.
